refactor model objects
