+++ app/connectors/GitHubConnector.scala
@@ -3,7 +3,7 @@ package connectors
 import cats.data.EitherT
 import cats.data.EitherT.rightT
 import com.google.inject._
-import models.{APIError, GitHubUser}
+import models.error._
 import play.api.libs.json.{JsError, JsSuccess, OFormat, Reads}
 import play.api.libs.ws.{WSClient, WSResponse}
 import scala.concurrent.{ExecutionContext, Future}
@@ -34,4 +34,26 @@ class GitHubConnector @Inject()(ws: WSClient) {
       }
     }
   }
+
+//  def put[Response](url: String)(implicit rds: Reads[Response], ec: ExecutionContext): EitherT[Future, APIError, Response] = {
+//
+//    val request = ws.url(url).addHttpHeaders(
+//      "Accept" -> "application/vnd.github+json",
+//      "Authorization" -> s"Bearer $personalAccessToken")
+//
+//    val response = request.put()
+//
+//    EitherT {
+//      response.map { result =>
+//        if (result.status == 200) {
+//          Right(result.json.as[Response])
+//        } else {
+//          Left(APIError.BadAPIResponse(result.status, result.statusText))
+//        }
+//      }.recover { case _: WSResponse =>
+//        Left(APIError.BadAPIResponse(500, "Could not connect to API."))
+//      }
+//    }
+//  }
+
 }
