+++ app/models/forms/UsernameSearch.scala
@@ -1,4 +1,4 @@
-package models
+package models.forms
 
 import play.api.data.Form
 import play.api.data.Forms.{mapping, nonEmptyText}
