+++ app/models/mongo/DataModel.scala
@@ -1,4 +1,4 @@
-package models
+package models.mongo
 
 import play.api.libs.json.{Json, OFormat}
 import play.api.data._
