+++ app/views/display/githubUser.scala.html
@@ -1,4 +1,4 @@
-@import models.GitHubUser.userForm
+@import models.github.GitHubUser
 @import helper._
 
 @(user: GitHubUser)(implicit request: RequestHeader, messages: Messages)
