+++ app/views/forms/searchUsername.scala.html
@@ -1,5 +1,5 @@
 @import helper._
-@import models.UsernameSearch
+@import models.forms.UsernameSearch
 @(usernameForm: Form[UsernameSearch])(implicit request:RequestHeader, messages:Messages)
