+++ app/views/repos/dirContent.scala.html
@@ -1,4 +1,4 @@
-@import models.RepoContentItem
+@import models.github.RepoContentItem
 
 @(username: String, repoName: String, contents: Seq[RepoContentItem])(implicit request: RequestHeader, messages: Messages)
