+++ app/views/repos/repos.scala.html
@@ -1,4 +1,4 @@
-@import models.Repository
+@import models.github.Repository
 
 @(username: String, repos: Seq[Repository])(implicit request: RequestHeader, messages: Messages)
